Add a WIP clean up hook
